calling checkSub()
hey thomas, can you look at handleinput.cpp from lines 38 to lines 47.
for checkSub(), it’d be ideal if it returned a bool directly to the
calling function at the if statement (line 38). do you think you can
set up the backend for that so it just *happens* (like connect the
signals and slots in the back)